+++ std/src/console/mod.rs
@@ -221,6 +221,16 @@ pub trait Console {
     ///
     fn write(&mut self, text: &str) -> io::Result<()>;
 
+    /// Draws the outline of a circle at `_center` with `_radius` using the current drawing color.
+    fn draw_circle(&mut self, _center: PixelsXY, _radius: u16) -> io::Result<()> {
+        Err(io::Error::new(io::ErrorKind::Other, "No graphics support in this console"))
+    }
+
+    /// Draws a filled circle at `_center` with `_radius` using the current drawing color.
+    fn draw_circle_filled(&mut self, _center: PixelsXY, _radius: u16) -> io::Result<()> {
+        Err(io::Error::new(io::ErrorKind::Other, "No graphics support in this console"))
+    }
+
     /// Draws a line from `_x1y1` to `_x2y2` using the current drawing color.
     fn draw_line(&mut self, _x1y1: PixelsXY, _x2y2: PixelsXY) -> io::Result<()> {
         Err(io::Error::new(io::ErrorKind::Other, "No graphics support in this console"))